+++ setup.cfg
@@ -51,6 +51,7 @@ install_requires =
     importlib-metadata; python_version<"3.8"
     pandas
     biocframe>=0.3.7
+    biocutils>=0.0.1
     numpy
     prettytable
